+++ src/css_styles/AddTodo.css
@@ -1,0 +1,13 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.date-input {
+  margin-top: 10px;
+}
+
+.button-container {
+  margin-top: 20px; 
+}
