+++ src/css_styles/Todo.css
@@ -1,4 +1,3 @@
-
 .todo-container {
   list-style-type: none;
   text-align: center;
@@ -65,19 +64,19 @@
 }
 
 .action-button {
-  background-color: #007bff; 
-  border: none; 
-  color: #ffffff; 
-  padding: 8px 16px;
-  margin-left: 10px;
-  cursor: pointer; 
-  text-decoration: none;
-  border-radius: 5px; 
+  background-color: #ff6666;
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  border-radius: 5px;
+  cursor: pointer;
+
 }
 
 .action-button:hover {
-  background-color: #0056b3; 
+  background-color: #ff4d4d;
 }
+
 
 
 .viewList {
@@ -92,16 +91,29 @@
 }
 
 .todo-list.completed {
-  margin-right: 20px; /* Adjust spacing between the lists */
+  margin-right: 20px; 
 }
 
 .todo-list.incompleted {
-  margin-left: 20px; /* Adjust spacing between the lists */
+  margin-left: 20px; 
 }
 
 .todo-item {
-  margin-bottom: 10px; /* Adjust spacing between todo items */
+  margin-bottom: 10px; 
 }
+
+.todo-task {
+  margin-bottom: 5px; /* Adjust as needed */
+  text-align: center; /* Center align the task */
+}
+
+.todo-date {
+  margin-top: 5px; /* Add some space between task and date */
+  text-align: center; /* Center align the date */
+  font-size: inherit; /* Inherit font size from parent */
+  font-family: inherit; /* Inherit font family from parent */
+}
+
 
 .todo-card {
   border: 1px solid #ccc;
@@ -121,18 +133,3 @@
   margin: 0;
   cursor: pointer;
 }
-
-.todo-card button {
-  background-color: #ff6666;
-  color: white;
-  border: none;
-  padding: 5px 10px;
-  border-radius: 5px;
-  cursor: pointer;
-
-}
-
-.todo-card button:hover {
-  background-color: #ff4d4d;
-}
-
